09-cell-vision/pyproject.toml
[project]
name = "cell-vision"
version = "0.1.0"
description = "CellVision - Microscopy Image Cell Type Classifier using deep learning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Danny", email = "danny@example.com"},
]
keywords = ["microscopy", "cell-classification", "deep-learning", "pytorch", "biology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.6.0",
    "streamlit>=1.31.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cell_vision"]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
